refactor: renames <ResetPassword/> page-component, test files and improves its compositions
