+++ template/src/pages/reset-password/__tests__/ResetPassword.test.tsx
@@ -1,30 +1,33 @@
 import userEvent from '@testing-library/user-event';
-import ResetPassword from '../index';
+import ResetPassword from '../ResetPassword';
 import { renderWithRouter, screen, waitFor } from '../../../utils/test';
 import { makeServer } from '../../../server';
 
-let server: any;
-
-beforeEach(() => {
-  server = makeServer();
-});
-
-afterEach(() => {
-  server.shutdown();
-});
-
 describe('<ResetPassword />', () => {
-  it('should render the ResetPassword form UI', () => {
+  // eslint-disable-next-line
+  let server: any;
+
+  beforeEach(() => {
+    server = makeServer();
+  });
+
+  afterEach(() => {
+    server.shutdown();
+  });
+
+  it('should render the ResetPassword form UI', async () => {
     renderWithRouter(<ResetPassword />);
 
-    expect(
-      screen.getByText(/Fill the form to reset your password/i)
-    ).toBeInTheDocument();
-    expect(screen.getByLabelText('password')).toBeInTheDocument();
-    expect(screen.getByLabelText('Confirm password')).toBeInTheDocument();
-    expect(
-      screen.getByRole('button', { name: /Change password/i })
-    ).toBeInTheDocument();
+    await waitFor(() => {
+      expect(
+        screen.getByText(/Fill the form to reset your password/i)
+      ).toBeInTheDocument();
+      expect(screen.getByLabelText('password')).toBeInTheDocument();
+      expect(screen.getByLabelText('Confirm password')).toBeInTheDocument();
+      expect(
+        screen.getByRole('button', { name: /Change password/i })
+      ).toBeInTheDocument();
+    });
   });
 
   it('should render error messages when inputs are invalid', async () => {
@@ -41,7 +44,7 @@ describe('<ResetPassword />', () => {
 
   it('should fail password change due to incorrect id', async () => {
     renderWithRouter(<ResetPassword />, {
-      route: 'reset-password/incorrect_id',
+      route: '/reset-password/incorrect_id',
     });
     server.create('user', {
       username: 'user@t10.digital',
@@ -51,8 +54,9 @@ describe('<ResetPassword />', () => {
     userEvent.type(screen.getByLabelText('password'), '111111111111');
     userEvent.type(screen.getByLabelText('Confirm password'), '111111111111');
     userEvent.click(screen.getByTestId('submit-reset'));
+
     await waitFor(() => {
-      expect(screen.getByText(/Something went wrong/i)).toBeInTheDocument();
+      expect(screen.getByText(/user not found/i)).toBeInTheDocument();
     });
   });
 });
